third_party/blink/web_tests/http/tests/websocket/handshake-fail-by-404-expected.txt

Connection should fail immediately if the server returns a 404 error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
